requirements.txt
opencv-python==4.7.0.72
mediapipe==0.10.0
numpy==1.24.2
customtkinter==5.1.2
Pillow==9.5.0
screen-brightness-control==0.9.0
pycaw==2023.2.2
comtypes==1.2.0